@@ -53,6 +53,7 @@ import me.aap.utils.vfs.VirtualFileSystem;
 import me.aap.utils.vfs.VirtualFolder;
 import me.aap.utils.vfs.VirtualResource;
 import me.aap.utils.vfs.content.ContentFileSystem;
+import me.aap.utils.vfs.content.ContentFolder;
 import me.aap.utils.vfs.local.LocalFileSystem;
 
 /**
@@ -71,8 +72,8 @@ public class MetadataRetriever implements Closeable {
 	private static final String COL_ART = "Art";
 	private static final String COL_DURATION = "Duration";
 	private static final String COL_ID_PATTERN = COL_ID + " LIKE ? AND NOT " + COL_ID + " LIKE ?";
-	private static final String[] QUERY_COLUMNS = {COL_ID, COL_TITLE, COL_ALBUM, COL_ARTIST,
-			COL_DURATION, COL_ART};
+	private static final String[] QUERY_COLUMNS =
+			{COL_ID, COL_TITLE, COL_ALBUM, COL_ARTIST, COL_DURATION, COL_ART};
 	private static final byte ART_URI = 0;
 	private static final byte ART_HASH = 1;
 	private static final String[] CONTENT_COLUMNS;
@@ -80,24 +81,14 @@ public class MetadataRetriever implements Closeable {
 
 	static {
 		if (SDK_INT >= VERSION_CODES.R) {
-			CONTENT_COLUMNS = new String[]{
-					MediaStore.MediaColumns._ID,
-					MediaStore.MediaColumns.TITLE,
-					MediaStore.Audio.AudioColumns.DURATION,
-					MediaStore.Audio.AudioColumns.ARTIST,
-					MediaStore.Audio.AudioColumns.ALBUM,
-					MediaStore.Audio.AudioColumns.ALBUM_ARTIST,
-					MediaStore.Audio.AudioColumns.COMPOSER,
-					MediaStore.Audio.AudioColumns.GENRE
-			};
+			CONTENT_COLUMNS = new String[]{MediaStore.MediaColumns._ID, MediaStore.MediaColumns.TITLE,
+					MediaStore.Audio.AudioColumns.DURATION, MediaStore.Audio.AudioColumns.ARTIST,
+					MediaStore.Audio.AudioColumns.ALBUM, MediaStore.Audio.AudioColumns.ALBUM_ARTIST,
+					MediaStore.Audio.AudioColumns.COMPOSER, MediaStore.Audio.AudioColumns.GENRE};
 		} else {
-			CONTENT_COLUMNS = new String[]{
-					MediaStore.MediaColumns._ID,
-					MediaStore.MediaColumns.TITLE,
-					MediaStore.Audio.AudioColumns.DURATION,
-					MediaStore.Audio.AudioColumns.ARTIST,
-					MediaStore.Audio.AudioColumns.ALBUM
-			};
+			CONTENT_COLUMNS = new String[]{MediaStore.MediaColumns._ID, MediaStore.MediaColumns.TITLE,
+					MediaStore.Audio.AudioColumns.DURATION, MediaStore.Audio.AudioColumns.ARTIST,
+					MediaStore.Audio.AudioColumns.ALBUM};
 		}
 
 		CONTENT_COLUMNS_DATA = Arrays.copyOf(CONTENT_COLUMNS, CONTENT_COLUMNS.length + 1);
@@ -156,7 +147,7 @@ public class MetadataRetriever implements Closeable {
 		VirtualResource res = item.getResource();
 		VirtualFileSystem fs = res.getVirtualFileSystem();
 
-		if ((fs instanceof LocalFileSystem)) {
+		if (fs instanceof LocalFileSystem) {
 			try {
 				if (queryMediaStore(item, mb)) {
 					insertMetadata(mb, item);
@@ -165,7 +156,7 @@ public class MetadataRetriever implements Closeable {
 			} catch (Throwable ex) {
 				Log.e(ex, "Failed retrieve data from MediaStore: ", res);
 			}
-		} else if (fs instanceof ContentFileSystem) {
+		} else if ((fs instanceof ContentFileSystem) || "content".equals(res.getRid().getScheme())) {
 			try {
 				if (queryContentProvider(item, mb)) {
 					insertMetadata(mb, item);
@@ -176,8 +167,8 @@ public class MetadataRetriever implements Closeable {
 			}
 		}
 
-		int scanner = (mgr.vlcPlayer == null) ? MEDIA_SCANNER_DEFAULT
-				: item.getLib().getPrefs().getMediaScannerPref();
+		int scanner = (mgr.vlcPlayer == null) ? MEDIA_SCANNER_DEFAULT :
+				item.getLib().getPrefs().getMediaScannerPref();
 
 		switch (scanner) {
 			case MEDIA_SCANNER_DEFAULT:
@@ -198,8 +189,10 @@ public class MetadataRetriever implements Closeable {
 	}
 
 	private boolean queryContentProvider(PlayableItem item, MetaBuilder mb) {
-		Uri uri = item.getResource().getRid().toAndroidUri();
 		ContentResolver cr = App.get().getContentResolver();
+		Uri uri = item.getResource().getRid().toAndroidUri();
+		Uri mu = toMediaUri(uri);
+		if (mu != null) uri = mu;
 
 		try (Cursor c = cr.query(uri, CONTENT_COLUMNS, null, null, null)) {
 			if ((c == null) || !c.moveToFirst() || !addFields(c, mb)) return false;
@@ -216,9 +209,8 @@ public class MetadataRetriever implements Closeable {
 		String canonical = null;
 		ContentResolver cr = App.get().getContentResolver();
 		boolean video = item.isVideo();
-		Uri uri = video
-				? MediaStore.Video.Media.getContentUri("external")
-				: MediaStore.Audio.Media.getContentUri("external");
+		Uri uri = video ? MediaStore.Video.Media.getContentUri("external") :
+				MediaStore.Audio.Media.getContentUri("external");
 
 		try {
 			canonical = new File(path).getCanonicalPath();
@@ -248,11 +240,22 @@ public class MetadataRetriever implements Closeable {
 	private Map<String, MetadataBuilder> queryMediaStore(BrowsableItem item) {
 		if (!(item instanceof FolderItem)) return emptyMap();
 		VirtualResource r = item.getResource();
-		if (!(r.getVirtualFileSystem() instanceof LocalFileSystem) || !(r instanceof VirtualFolder)) {
-			return emptyMap();
+		String path = null;
+
+		if (r instanceof ContentFolder) {
+			if (SDK_INT < VERSION_CODES.Q) return emptyMap();
+			Uri uri = toMediaUri(r.getRid().toAndroidUri());
+			if (uri == null) return emptyMap();
+
+			try (Cursor c = App.get().getContentResolver()
+					.query(uri, new String[]{"_data"}, null, null, null)) {
+				if ((c != null) && c.moveToNext()) path = c.getString(0);
+			}
+		} else if ((r.getVirtualFileSystem() instanceof LocalFileSystem) &&
+				(r instanceof VirtualFolder)) {
+			path = item.getResource().getRid().getPath();
 		}
 
-		String path = item.getResource().getRid().getPath();
 		if (path == null) return emptyMap();
 		String dataQuery;
 		String[] dataArgs;
@@ -261,9 +264,11 @@ public class MetadataRetriever implements Closeable {
 		Map<String, MetadataBuilder> m = new HashMap<>();
 		ContentResolver cr = App.get().getContentResolver();
 
-		try {
-			canonical = new File(path).getCanonicalPath();
-		} catch (Throwable ignore) {
+		if (!(r instanceof ContentFolder)) {
+			try {
+				canonical = new File(path).getCanonicalPath();
+			} catch (Throwable ignore) {
+			}
 		}
 
 		if (canonical != null) {
@@ -275,8 +280,10 @@ public class MetadataRetriever implements Closeable {
 		}
 
 		id = FileItem.SCHEME + id.substring(FolderItem.SCHEME.length());
-		queryMediaStore(cr, MediaStore.Audio.Media.getContentUri("external"), dataQuery, dataArgs, id, m);
-		queryMediaStore(cr, MediaStore.Video.Media.getContentUri("external"), dataQuery, dataArgs, id, m);
+		queryMediaStore(cr, MediaStore.Audio.Media.getContentUri("external"), dataQuery, dataArgs, id,
+				m);
+		queryMediaStore(cr, MediaStore.Video.Media.getContentUri("external"), dataQuery, dataArgs, id,
+				m);
 		return m;
 	}
 
@@ -301,7 +308,18 @@ public class MetadataRetriever implements Closeable {
 		}
 	}
 
-	private boolean addFields(Cursor c, MetaBuilder mb) {
+	private static Uri toMediaUri(Uri uri) {
+		if (SDK_INT >= VERSION_CODES.Q) {
+			try {
+				return MediaStore.getMediaUri(App.get(), uri);
+			} catch (Exception ex) {
+				Log.d(ex, "Failed to get media uri");
+			}
+		}
+		return null;
+	}
+
+	private static boolean addFields(Cursor c, MetaBuilder mb) {
 		try {
 			long dur = c.getLong(2);
 			if (dur <= 0) return false;
@@ -330,7 +348,8 @@ public class MetadataRetriever implements Closeable {
 		return true;
 	}
 
-	public FutureSupplier<Map<String, MetadataBuilder>> queryMetadata(String idPattern, BrowsableItem br) {
+	public FutureSupplier<Map<String, MetadataBuilder>> queryMetadata(String idPattern,
+																																		BrowsableItem br) {
 		if (db == null) return queue.enqueue(() -> queryMediaStore(br));
 		return queue.enqueue(() -> {
 			Map<String, MetadataBuilder> m = query(idPattern);
@@ -345,16 +364,16 @@ public class MetadataRetriever implements Closeable {
 	public FutureSupplier<List<String>> queryIds(String pattern, int max) {
 		return (db != null) ? queue.enqueue(() -> {
 			List<String> ids = new ArrayList<>(max);
-			try (Cursor c = db.query(TABLE, new String[]{COL_ID}, COL_TITLE + " = ? OR " +
-							COL_ARTIST + " = ? OR " + COL_ALBUM + " = ? LIMIT " + max,
+			try (Cursor c = db.query(TABLE, new String[]{COL_ID},
+					COL_TITLE + " = ? OR " + COL_ARTIST + " = ? OR " + COL_ALBUM + " = ? LIMIT " + max,
 					new String[]{pattern, pattern, pattern}, null, null, null)) {
 				while (c.moveToNext()) ids.add(c.getString(0));
 			}
 			if (!ids.isEmpty()) return ids;
 
 			String[] p = {'%' + pattern + '%'};
-			try (Cursor c = db.query(TABLE, new String[]{COL_ID}, COL_TITLE + " LIKE ?  LIMIT " + max,
-					p, null, null, null)) {
+			try (Cursor c = db.query(TABLE, new String[]{COL_ID}, COL_TITLE + " LIKE ?  LIMIT " + max, p,
+					null, null, null)) {
 				while (c.moveToNext()) ids.add(c.getString(0));
 			}
 			if (!ids.isEmpty()) return ids;
@@ -363,8 +382,8 @@ public class MetadataRetriever implements Closeable {
 				while (c.moveToNext()) ids.add(c.getString(0));
 			}
 			if (!ids.isEmpty()) return ids;
-			try (Cursor c = db.query(TABLE, new String[]{COL_ID}, COL_ALBUM + " LIKE ?  LIMIT " + max,
-					p, null, null, null)) {
+			try (Cursor c = db.query(TABLE, new String[]{COL_ID}, COL_ALBUM + " LIKE ?  LIMIT " + max, p,
+					null, null, null)) {
 				while (c.moveToNext()) ids.add(c.getString(0));
 			}
 			return ids.isEmpty() ? emptyList() : ids;
@@ -425,9 +444,8 @@ public class MetadataRetriever implements Closeable {
 	private MetadataBuilder queryMetadata(PlayableItem item) {
 		if (db == null) return null;
 
-		try (Cursor c = db.query(TABLE, QUERY_COLUMNS, COL_ID + " = ?",
-				new String[]{item.getOrigId()}, null, null, null);
-				 SharedTextBuilder tb = SharedTextBuilder.get()) {
+		try (Cursor c = db.query(TABLE, QUERY_COLUMNS, COL_ID + " = ?", new String[]{item.getOrigId()},
+				null, null, null); SharedTextBuilder tb = SharedTextBuilder.get()) {
 			if (!c.moveToNext()) return null;
 			MetadataBuilder meta = new MetaBuilder();
 			readMetadata(meta, c, tb);
@@ -509,18 +527,12 @@ public class MetadataRetriever implements Closeable {
 			ps.applyIntPref(version, BuildConfig.VERSION_CODE);
 		}
 
-		db.execSQL("CREATE TABLE IF NOT EXISTS " + TABLE + "(" +
-				COL_ID + " VARCHAR NOT NULL UNIQUE, " +
-				COL_TITLE + " VARCHAR, " +
-				COL_ALBUM + " VARCHAR, " +
-				COL_ARTIST + " VARCHAR, " +
-				COL_ALBUM_ARTIST + " VARCHAR, " +
-				COL_COMPOSER + " VARCHAR, " +
-				COL_WRITER + " VARCHAR, " +
-				COL_GENRE + " VARCHAR, " +
-				COL_DURATION + " INTEGER, " +
-				COL_ART + " BLOB " +
-				");");
+		db.execSQL(
+				"CREATE TABLE IF NOT EXISTS " + TABLE + "(" + COL_ID + " VARCHAR NOT NULL UNIQUE," + " " +
+						COL_TITLE + " VARCHAR, " + COL_ALBUM + " VARCHAR, " + COL_ARTIST + " VARCHAR, " +
+						COL_ALBUM_ARTIST + " VARCHAR, " + COL_COMPOSER + " VARCHAR, " + COL_WRITER +
+						" VARCHAR, " + COL_GENRE + " VARCHAR, " + COL_DURATION + " INTEGER, " + COL_ART +
+						" BLOB " + ");");
 	}
 
 	private static final class MetaBuilder extends MetadataBuilder {
@@ -531,30 +543,17 @@ public class MetadataRetriever implements Closeable {
 		@Override
 		public void putString(String k, String v) {
 			switch (k) {
-				case MediaMetadataCompat.METADATA_KEY_TITLE:
-					values.put(COL_TITLE, v);
-					break;
-				case MediaMetadataCompat.METADATA_KEY_ALBUM:
-					values.put(COL_ALBUM, v);
-					break;
-				case MediaMetadataCompat.METADATA_KEY_ARTIST:
-					values.put(COL_ARTIST, v);
-					break;
-				case MediaMetadataCompat.METADATA_KEY_ALBUM_ARTIST:
-					values.put(COL_ALBUM_ARTIST, v);
-					break;
-				case MediaMetadataCompat.METADATA_KEY_COMPOSER:
-					values.put(COL_COMPOSER, v);
-					break;
-				case MediaMetadataCompat.METADATA_KEY_WRITER:
-					values.put(COL_WRITER, v);
-					break;
-				case MediaMetadataCompat.METADATA_KEY_GENRE:
-					values.put(COL_GENRE, v);
-					break;
-				case MediaMetadataCompat.METADATA_KEY_ALBUM_ART_URI:
+				case MediaMetadataCompat.METADATA_KEY_TITLE -> values.put(COL_TITLE, v);
+				case MediaMetadataCompat.METADATA_KEY_ALBUM -> values.put(COL_ALBUM, v);
+				case MediaMetadataCompat.METADATA_KEY_ARTIST -> values.put(COL_ARTIST, v);
+				case MediaMetadataCompat.METADATA_KEY_ALBUM_ARTIST -> values.put(COL_ALBUM_ARTIST, v);
+				case MediaMetadataCompat.METADATA_KEY_COMPOSER -> values.put(COL_COMPOSER, v);
+				case MediaMetadataCompat.METADATA_KEY_WRITER -> values.put(COL_WRITER, v);
+				case MediaMetadataCompat.METADATA_KEY_GENRE -> values.put(COL_GENRE, v);
+				case MediaMetadataCompat.METADATA_KEY_ALBUM_ART_URI -> {
 					setImageUri(v);
 					return;
+				}
 			}
 
 			super.putString(k, v);
