@@ -2,9 +2,6 @@ package me.aap.fermata.ui.activity;
 
 import static me.aap.fermata.BuildConfig.AUTO;
 
-import android.content.Context;
-import android.content.res.Configuration;
-
 import androidx.annotation.Nullable;
 
 import java.util.List;
@@ -22,7 +19,8 @@ import me.aap.utils.ui.view.NavBarView;
 /**
  * @author Andrey Pavlenko
  */
-public interface MainActivityPrefs extends SharedPreferenceStore, EventBroadcaster<PreferenceStore.Listener> {
+public interface MainActivityPrefs
+		extends SharedPreferenceStore, EventBroadcaster<PreferenceStore.Listener> {
 	int THEME_DARK = 0;
 	int THEME_LIGHT = 1;
 	int THEME_DAY_NIGHT = 2;
@@ -37,10 +35,12 @@ public interface MainActivityPrefs extends SharedPreferenceStore, EventBroadcast
 	int LOCALE_IT = 2;
 	int LOCALE_TR = 3;
 	int LOCALE_DE = 4;
+	int LOCALE_PT = 5;
 	Pref<IntSupplier> THEME_MAIN = Pref.i("THEME_MAIN", THEME_DARK);
 	Pref<BooleanSupplier> HIDE_BARS = Pref.b("HIDE_BARS", false);
 	Pref<BooleanSupplier> FULLSCREEN = Pref.b("FULLSCREEN", false);
 	Pref<BooleanSupplier> SHOW_PG_UP_DOWN = Pref.b("SHOW_PG_UP_DOWN", true);
+	Pref<BooleanSupplier> USE_DPAD_CURSOR = AUTO ? Pref.b("USE_DPAD_CURSOR", true) : null;
 	Pref<IntSupplier> NAV_BAR_POS = Pref.i("NAV_BAR_POS", NavBarView.POSITION_BOTTOM);
 	Pref<DoubleSupplier> NAV_BAR_SIZE = Pref.f("NAV_BAR_SIZE", 1f);
 	Pref<DoubleSupplier> TOOL_BAR_SIZE = Pref.f("TOOL_BAR_SIZE", 1f);
@@ -49,6 +49,8 @@ public interface MainActivityPrefs extends SharedPreferenceStore, EventBroadcast
 	Pref<BooleanSupplier> GRID_VIEW = Pref.b("GRID_VIEW", false);
 	Pref<DoubleSupplier> P_SPLIT_PERCENT = Pref.f("P_SPLIT_PERCENT", 0.6f);
 	Pref<DoubleSupplier> L_SPLIT_PERCENT = Pref.f("L_SPLIT_PERCENT", 0.4f);
+	Pref<DoubleSupplier> P_SPLIT_PERCENT_SUB = Pref.f("P_SPLIT_PERCENT_SUB", 0.5f);
+	Pref<DoubleSupplier> L_SPLIT_PERCENT_SUB = Pref.f("L_SPLIT_PERCENT_SUB", 0.5f);
 	Pref<Supplier<String>> SHOW_ADDON_ON_START = Pref.s("SHOW_ADDON_ON_START", (String) null);
 	Pref<BooleanSupplier> CHECK_UPDATES = Pref.b("CHECK_UPDATES", true);
 	Pref<BooleanSupplier> SYS_BARS_ON_VIDEO_TOUCH = Pref.b("SYS_BARS_ON_VIDEO_TOUCH", false);
@@ -57,29 +59,25 @@ public interface MainActivityPrefs extends SharedPreferenceStore, EventBroadcast
 	Pref<IntSupplier> BRIGHTNESS = Pref.i("BRIGHTNESS", 255);
 	Pref<BooleanSupplier> VOICE_CONTROl_ENABLED = Pref.b("VOICE_CONTROl_ENABLED", false);
 	Pref<BooleanSupplier> VOICE_CONTROl_FB = Pref.b("VOICE_CONTROl_FB", false);
-	Pref<BooleanSupplier> VOICE_CONTROl_M = Pref.b("VOICE_CONTROl_M", false);
 	Pref<Supplier<String>> VOICE_CONTROL_SUBST = Pref.s("VOICE_CONTROL_SUBST", "");
+	Pref<Supplier<String>> VOICE_CONTROL_LANG =
+			Pref.s("VOICE_CONTROL_LANG", () -> Locale.getDefault().toLanguageTag());
 	Pref<IntSupplier> CLOCK_POS = Pref.i("CLOCK_POS", CLOCK_POS_NONE);
-	Pref<IntSupplier> LOCALE = Pref.i("LOCALE", () -> {
-		switch (Locale.getDefault().getLanguage()) {
-			case "ru":
-				return LOCALE_RU;
-			case "it":
-				return LOCALE_IT;
-			case "tr":
-				return LOCALE_TR;
-			case "de":
-				return LOCALE_DE;
-			default:
-				return LOCALE_EN;
-		}
+	Pref<IntSupplier> LOCALE = Pref.i("LOCALE", () -> switch (Locale.getDefault().getLanguage()) {
+		case "ru" -> LOCALE_RU;
+		case "it" -> LOCALE_IT;
+		case "tr" -> LOCALE_TR;
+		case "de" -> LOCALE_DE;
+		case "pt" -> LOCALE_PT;
+		default -> LOCALE_EN;
 	});
 
 	Pref<IntSupplier> THEME_AA = Pref.i("THEME_AA", THEME_DARK);
 	Pref<BooleanSupplier> HIDE_BARS_AA = AUTO ? Pref.b("HIDE_BARS_AA", false) : null;
 	Pref<BooleanSupplier> FULLSCREEN_AA = AUTO ? Pref.b("FULLSCREEN_AA", false) : null;
 	Pref<BooleanSupplier> SHOW_PG_UP_DOWN_AA = AUTO ? Pref.b("SHOW_PG_UP_DOWN_AA", true) : null;
-	Pref<IntSupplier> NAV_BAR_POS_AA = AUTO ? Pref.i("NAV_BAR_POS_AA", NavBarView.POSITION_BOTTOM) : null;
+	Pref<IntSupplier> NAV_BAR_POS_AA =
+			AUTO ? Pref.i("NAV_BAR_POS_AA", NavBarView.POSITION_BOTTOM) : null;
 	Pref<DoubleSupplier> NAV_BAR_SIZE_AA = AUTO ? Pref.f("NAV_BAR_SIZE_AA", 1f) : null;
 	Pref<DoubleSupplier> TOOL_BAR_SIZE_AA = AUTO ? Pref.f("TOOL_BAR_SIZE_AA", 1f) : null;
 	Pref<DoubleSupplier> CONTROL_PANEL_SIZE_AA = AUTO ? Pref.f("CONTROL_PANEL_SIZE_AA", 1f) : null;
@@ -97,22 +95,6 @@ public interface MainActivityPrefs extends SharedPreferenceStore, EventBroadcast
 
 	default int getThemePref(boolean auto) {
 		return (AUTO && auto) ? getIntPref(THEME_AA) : getIntPref(THEME_MAIN);
-	}
-
-	default float getSplitPercent(Context ctx) {
-		if (ctx.getResources().getConfiguration().orientation == Configuration.ORIENTATION_PORTRAIT) {
-			return getFloatPref(P_SPLIT_PERCENT);
-		} else {
-			return getFloatPref(L_SPLIT_PERCENT);
-		}
-	}
-
-	default void setSplitPercent(Context ctx, float percent) {
-		if (ctx.getResources().getConfiguration().orientation == Configuration.ORIENTATION_PORTRAIT) {
-			applyFloatPref(P_SPLIT_PERCENT, percent);
-		} else {
-			applyFloatPref(L_SPLIT_PERCENT, percent);
-		}
 	}
 
 	@Nullable
@@ -151,6 +133,10 @@ public interface MainActivityPrefs extends SharedPreferenceStore, EventBroadcast
 	default boolean getShowPgUpDownPref(MainActivityDelegate a) {
 		if (AUTO && a.isCarActivity()) return getBooleanPref(SHOW_PG_UP_DOWN_AA);
 		return getBooleanPref(SHOW_PG_UP_DOWN);
+	}
+
+	default boolean useDpadCursor(MainActivityDelegate a) {
+		return AUTO && a.isCarActivity() && getBooleanPref(USE_DPAD_CURSOR);
 	}
 
 	static boolean hasNavBarPosPref(MainActivityDelegate a, List<Pref<?>> prefs) {
@@ -244,8 +230,8 @@ public interface MainActivityPrefs extends SharedPreferenceStore, EventBroadcast
 		return getBooleanPref(VOICE_CONTROl_FB);
 	}
 
-	default boolean getVoiceControlMenuPref() {
-		return getBooleanPref(VOICE_CONTROl_M);
+	default String getVoiceControlLang(MainActivityDelegate a) {
+		return a.getPrefs().getStringPref(VOICE_CONTROL_LANG);
 	}
 
 	default int getClockPosPref() {
@@ -253,17 +239,13 @@ public interface MainActivityPrefs extends SharedPreferenceStore, EventBroadcast
 	}
 
 	default Locale getLocalePref() {
-		switch (getIntPref(LOCALE)) {
-			case LOCALE_RU:
-				return new Locale("ru");
-			case LOCALE_IT:
-				return Locale.ITALIAN;
-			case LOCALE_TR:
-				return new Locale("tr");
-			case LOCALE_DE:
-				return new Locale("de");
-			default:
-				return Locale.ENGLISH;
-		}
+		return switch (getIntPref(LOCALE)) {
+			case LOCALE_RU -> new Locale("ru");
+			case LOCALE_IT -> Locale.ITALIAN;
+			case LOCALE_TR -> new Locale("tr");
+			case LOCALE_DE -> new Locale("de");
+			case LOCALE_PT -> new Locale("pt");
+			default -> Locale.ENGLISH;
+		};
 	}
 }
