@@ -1,63 +1,70 @@
 package me.aap.fermata.ui.view;
 
-import android.annotation.SuppressLint;
+import static me.aap.fermata.R.id.subtitles_fragment;
+import static me.aap.fermata.ui.activity.MainActivityPrefs.L_SPLIT_PERCENT;
+import static me.aap.fermata.ui.activity.MainActivityPrefs.P_SPLIT_PERCENT;
+
 import android.content.Context;
 import android.content.res.Configuration;
 import android.util.AttributeSet;
-import android.view.MotionEvent;
 import android.view.View;
 import android.widget.Toast;
 
 import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
-import androidx.appcompat.widget.AppCompatImageView;
 import androidx.constraintlayout.widget.ConstraintLayout;
-import androidx.constraintlayout.widget.ConstraintSet;
 import androidx.constraintlayout.widget.Guideline;
 import androidx.swiperefreshlayout.widget.SwipeRefreshLayout;
 
 import me.aap.fermata.R;
 import me.aap.fermata.media.engine.MediaEngine;
+import me.aap.fermata.media.engine.SubtitleStreamInfo;
 import me.aap.fermata.media.lib.MediaLib;
 import me.aap.fermata.media.service.FermataServiceUiBinder;
 import me.aap.fermata.media.service.MediaSessionCallback;
 import me.aap.fermata.ui.activity.MainActivityDelegate;
 import me.aap.fermata.ui.activity.MainActivityListener;
+import me.aap.fermata.ui.fragment.MediaLibFragment;
+import me.aap.fermata.ui.fragment.SubtitlesFragment;
 import me.aap.utils.app.App;
+import me.aap.utils.function.DoubleSupplier;
+import me.aap.utils.pref.PreferenceStore.Pref;
 import me.aap.utils.ui.fragment.ActivityFragment;
 
 /**
  * @author Andrey Pavlenko
  */
-public class BodyLayout extends ConstraintLayout implements
-		SwipeRefreshLayout.OnRefreshListener, SwipeRefreshLayout.OnChildScrollUpCallback,
-		View.OnTouchListener, FermataServiceUiBinder.Listener, MainActivityListener {
+public class BodyLayout extends SplitLayout
+		implements SwipeRefreshLayout.OnRefreshListener, SwipeRefreshLayout.OnChildScrollUpCallback,
+		MainActivityListener, FermataServiceUiBinder.Listener, MediaSessionCallback.Listener {
 	private Mode mode;
 
 	public BodyLayout(@NonNull Context ctx, @Nullable AttributeSet attrs) {
-		this(ctx, attrs, true);
-	}
-
-	@SuppressLint("ClickableViewAccessibility")
-	private BodyLayout(@NonNull Context ctx, @Nullable AttributeSet attrs, boolean init) {
 		super(ctx, attrs);
-		inflate(ctx, isPortrait() ? R.layout.body_layout : R.layout.body_layout_land, this);
-		if (!init) return;
 
-		MainActivityDelegate a = getActivity();
 		SwipeRefreshLayout srl = getSwipeRefresh();
 		srl.setId(R.id.swiperefresh);
 		srl.setOnRefreshListener(this);
 		srl.setOnChildScrollUpCallback(this);
-		getSplitLine().setOnTouchListener(this);
-		getSplitHandle().setOnTouchListener(this);
-		setSplitPercent(a.getPrefs().getSplitPercent(ctx));
 		setMode(Mode.FRAME);
 
-		FermataServiceUiBinder b = a.getMediaServiceBinder();
-		b.addBroadcastListener(this);
-		a.addBroadcastListener(this, FRAGMENT_CHANGED | ACTIVITY_DESTROY);
-		onPlayableChanged(null, b.getCurrentItem());
+		MainActivityDelegate.getActivityDelegate(ctx).onSuccess(a -> {
+			FermataServiceUiBinder b = a.getMediaServiceBinder();
+			b.addBroadcastListener(this);
+			a.addBroadcastListener(this, FRAGMENT_CHANGED | ACTIVITY_DESTROY);
+			b.getMediaSessionCallback().addBroadcastListener(this);
+			onPlayableChanged(null, b.getCurrentItem());
+		});
+	}
+
+	@Override
+	protected int getLayout(boolean portrait) {
+		return portrait ? R.layout.body_layout : R.layout.body_layout_land;
+	}
+
+	@Override
+	protected Pref<DoubleSupplier> getSplitPercentPref(boolean portrait) {
+		return portrait ? P_SPLIT_PERCENT : L_SPLIT_PERCENT;
 	}
 
 	public Mode getMode() {
@@ -84,15 +91,15 @@ public class BodyLayout extends ConstraintLayout implements
 		VideoView vv = getVideoView();
 
 		switch (mode) {
-			case FRAME:
+			case FRAME -> {
 				vv.setVisibility(GONE);
 				getSplitLine().setVisibility(GONE);
 				getSplitHandle().setVisibility(GONE);
 				getSwipeRefresh().setVisibility(VISIBLE);
 				lp.guidePercent = isPortrait() ? 0f : 1f;
 				a.setVideoMode(false, vv);
-				break;
-			case VIDEO:
+			}
+			case VIDEO -> {
 				vv.setVisibility(VISIBLE);
 				getSplitLine().setVisibility(GONE);
 				getSplitHandle().setVisibility(GONE);
@@ -101,50 +108,21 @@ public class BodyLayout extends ConstraintLayout implements
 				vv.showVideo(true);
 				a.setVideoMode(true, vv);
 				App.get().getHandler().post(vv::requestFocus);
-				break;
-			case BOTH:
+			}
+			case BOTH -> {
 				vv.setVisibility(VISIBLE);
 				getSplitLine().setVisibility(VISIBLE);
 				getSplitHandle().setVisibility(VISIBLE);
 				getSwipeRefresh().setVisibility(VISIBLE);
-				lp.guidePercent = a.getPrefs().getSplitPercent(getContext());
+				lp.guidePercent = a.getPrefs().getFloatPref(getSplitPercentPref(isPortrait()));
 				vv.showVideo(true);
 				a.setVideoMode(true, vv);
 				MediaItemListView.focusActive(getContext(), vv);
-				break;
+			}
 		}
 
 		gl.setLayoutParams(lp);
 		a.fireBroadcastEvent(MODE_CHANGED);
-	}
-
-	@Override
-	protected void onConfigurationChanged(Configuration newConfig) {
-		super.onConfigurationChanged(newConfig);
-		layout(newConfig.orientation);
-	}
-
-	private void layout(int orientation) {
-		if (orientation == Configuration.ORIENTATION_PORTRAIT) layoutPortrait();
-		else layoutLandscape();
-	}
-
-	private void layoutPortrait() {
-		ConstraintSet cs = new ConstraintSet();
-		cs.clone(new BodyLayout(getContext(), null, false));
-		cs.applyTo(this);
-		getSplitHandle().setImageResource(R.drawable.horizontal_split);
-		setSplitPercent(getActivity().getPrefs().getSplitPercent(getContext()));
-		setMode(getMode());
-	}
-
-	private void layoutLandscape() {
-		ConstraintSet cs = new ConstraintSet();
-		cs.clone(new BodyLayout(getContext(), null, false));
-		cs.applyTo(this);
-		getSplitHandle().setImageResource(R.drawable.vertical_split);
-		setSplitPercent(getActivity().getPrefs().getSplitPercent(getContext()));
-		setMode(getMode());
 	}
 
 	public VideoView getVideoView() {
@@ -155,22 +133,18 @@ public class BodyLayout extends ConstraintLayout implements
 		return findViewById(R.id.swiperefresh);
 	}
 
-	private Guideline getGuideline() {
-		return findViewById(R.id.guideline);
-	}
-
-	private View getSplitLine() {
-		return findViewById(R.id.split_line);
-	}
-
-	private AppCompatImageView getSplitHandle() {
-		return findViewById(R.id.split_handle);
+	@Override
+	protected void onConfigurationChanged(Configuration newConfig) {
+		super.onConfigurationChanged(newConfig);
+		setMode(getMode());
 	}
 
 	@Override
 	public void onActivityEvent(MainActivityDelegate a, long e) {
 		if (handleActivityDestroyEvent(a, e)) {
-			a.getMediaServiceBinder().removeBroadcastListener(this);
+			FermataServiceUiBinder b = a.getMediaServiceBinder();
+			b.removeBroadcastListener(this);
+			b.getMediaSessionCallback().removeBroadcastListener(this);
 		} else if (e == FRAGMENT_CHANGED) {
 			if (a.getActiveMediaLibFragment() == null) {
 				setMode(Mode.FRAME);
@@ -185,7 +159,8 @@ public class BodyLayout extends ConstraintLayout implements
 
 				MediaLib.PlayableItem i = eng.getSource();
 
-				if ((i != null) && i.isVideo() && !i.isExternal() && (cb.getVideoView() == getVideoView())) {
+				if ((i != null) && i.isVideo() && eng.isSplitModeSupported() &&
+						(cb.getVideoView() == getVideoView())) {
 					setMode(Mode.BOTH);
 				} else {
 					setMode(Mode.FRAME);
@@ -197,13 +172,37 @@ public class BodyLayout extends ConstraintLayout implements
 	@Override
 	public void onPlayableChanged(MediaLib.PlayableItem oldItem, MediaLib.PlayableItem newItem) {
 		MainActivityDelegate a = getActivity();
-		if (a.getActiveMediaLibFragment() == null) return;
+		var f = a.getActiveFragment();
+		if (f instanceof SubtitlesFragment) a.goToCurrent();
+		else if ((f != null) && !(f instanceof MediaLibFragment)) return;
+		MediaEngine eng = a.getMediaServiceBinder().getCurrentEngine();
 
-		if ((newItem == null) || !newItem.isVideo() || newItem.isExternal()) {
-			setMode(BodyLayout.Mode.FRAME);
+		if ((newItem == null) || !newItem.isVideo() || (eng == null) || !eng.isSplitModeSupported()) {
+			setMode(Mode.FRAME);
 		} else {
-			if (isFrameMode()) setMode(BodyLayout.Mode.VIDEO);
+			if (!eng.isVideoModeRequired()) setMode(Mode.FRAME);
+			else if (isFrameMode()) setMode(Mode.VIDEO);
 			else getVideoView().showVideo(false);
+		}
+
+		if ((eng != null) && (newItem != null) && !newItem.isVideo() && (getMode() == Mode.FRAME)) {
+			eng.selectSubtitleStream();
+		}
+	}
+
+	@Override
+	public void onSubtitleStreamChanged(MediaSessionCallback cb, @Nullable SubtitleStreamInfo info) {
+		if (getMode() != Mode.FRAME) return;
+		var i = cb.getCurrentItem();
+		if ((i == null) || i.isVideo()) return;
+		var a = getActivity();
+		var f = a.getActiveFragment();
+		if (info == null) {
+			if (f instanceof SubtitlesFragment) a.goToCurrent();
+		} else if (f instanceof SubtitlesFragment) {
+			((SubtitlesFragment) f).restart();
+		} else {
+			a.post(() -> a.showFragment(subtitles_fragment));
 		}
 	}
 
@@ -224,39 +223,6 @@ public class BodyLayout extends ConstraintLayout implements
 		if (a.isMenuActive()) return true;
 		ActivityFragment f = a.getActiveFragment();
 		return (f != null) && f.canScrollUp();
-	}
-
-	@Override
-	public boolean onTouch(View v, MotionEvent event) {
-		switch (event.getActionMasked()) {
-			case MotionEvent.ACTION_DOWN:
-				return true;
-			case MotionEvent.ACTION_MOVE:
-				float percent;
-				if (isPortrait()) percent = (event.getRawY()) / getRootView().getHeight();
-				else percent = (event.getRawX()) / getRootView().getWidth();
-
-				setSplitPercent(percent);
-				getActivity().getPrefs().setSplitPercent(getContext(), percent);
-				return true;
-			default:
-				return false;
-		}
-	}
-
-	private void setSplitPercent(float percent) {
-		Guideline gl = getGuideline();
-		ConstraintLayout.LayoutParams params = (ConstraintLayout.LayoutParams) gl.getLayoutParams();
-		params.guidePercent = Math.max(0.1f, Math.min(0.9f, percent));
-		gl.setLayoutParams(params);
-	}
-
-	private boolean isPortrait() {
-		return getResources().getConfiguration().orientation == Configuration.ORIENTATION_PORTRAIT;
-	}
-
-	private MainActivityDelegate getActivity() {
-		return MainActivityDelegate.get(getContext());
 	}
 
 	public enum Mode {
