@@ -9,7 +9,7 @@ import me.aap.utils.pref.PreferenceStore;
 /**
  * @author Andrey Pavlenko
  */
-public class AddonInfo {
+public class AddonInfo implements Comparable<AddonInfo> {
 	public final String moduleName;
 	public final String className;
 	@StringRes
@@ -17,13 +17,21 @@ public class AddonInfo {
 	@DrawableRes
 	public final int icon;
 	public final PreferenceStore.Pref<BooleanSupplier> enabledPref;
+	public final int order;
 
-	public AddonInfo(String moduleName, String className, int addonName, int icon) {
+	public AddonInfo(String moduleName, String className, int addonName, int icon,
+									 Integer order, Boolean enabled) {
 		this.moduleName = moduleName;
 		this.className = className;
 		this.addonName = addonName;
 		this.icon = icon;
-		enabledPref = PreferenceStore.Pref.b(className + "_enabled", this::isInstalled);
+		this.order = (order == null) ? 0 : order;
+		enabledPref = PreferenceStore.Pref.b(className + "_enabled",
+				(enabled == null) ? this::isInstalled : () -> enabled);
+	}
+
+	public String getModuleName() {
+		return moduleName;
 	}
 
 	public boolean isInstalled() {
@@ -33,5 +41,10 @@ public class AddonInfo {
 		} catch (Exception ignore) {
 			return false;
 		}
+	}
+
+	@Override
+	public int compareTo(AddonInfo ai) {
+		return Integer.compare(order, ai.order);
 	}
 }
