@@ -54,8 +54,15 @@ public class AddonManager extends BasicEventBroadcaster<AddonManager.Listener>
 	}
 
 	@Nullable
-	public FermataAddon getAddon(String className) {
-		return addons.get(className);
+	public FermataAddon getAddon(String moduleOrClassName) {
+		if (moduleOrClassName.indexOf('.') < 0) {
+			for (FermataAddon a : addons.values()) {
+				if (a.getInfo().getModuleName().equals(moduleOrClassName)) return a;
+			}
+		} else {
+			return addons.get(moduleOrClassName);
+		}
+		return null;
 	}
 
 	@SuppressWarnings("unchecked")
@@ -78,7 +85,9 @@ public class AddonManager extends BasicEventBroadcaster<AddonManager.Listener>
 	@Nullable
 	public ActivityFragment createFragment(@IdRes int id) {
 		for (FermataAddon a : getAddons()) {
-			if (a.getAddonId() == id) return a.createFragment();
+			if (a instanceof FermataFragmentAddon) {
+				if (a.getAddonId() == id) return ((FermataFragmentAddon) a).createFragment();
+			}
 		}
 		return null;
 	}
