@@ -1,13 +1,15 @@
 package me.aap.fermata.addon;
 
+import android.content.Intent;
+
 import androidx.annotation.IdRes;
 import androidx.annotation.NonNull;
 
 import me.aap.fermata.BuildConfig;
+import me.aap.fermata.ui.activity.MainActivityDelegate;
 import me.aap.utils.misc.ChangeableCondition;
 import me.aap.utils.pref.PreferenceSet;
 import me.aap.utils.pref.PreferenceStore;
-import me.aap.utils.ui.fragment.ActivityFragment;
 
 /**
  * @author Andrey Pavlenko
@@ -20,13 +22,6 @@ public interface FermataAddon {
 	@NonNull
 	AddonInfo getInfo();
 
-	@NonNull
-	ActivityFragment createFragment();
-
-	default int getFragmentId() {
-		return getAddonId();
-	}
-
 	default void contributeSettings(PreferenceStore store, PreferenceSet set, ChangeableCondition visibility) {
 	}
 
@@ -34,6 +29,10 @@ public interface FermataAddon {
 	}
 
 	default void uninstall() {
+	}
+
+	default boolean handleIntent(MainActivityDelegate a, Intent intent) {
+		return false;
 	}
 
 	@NonNull
