@@ -3,21 +3,24 @@ package me.aap.fermata.media.engine;
 import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
 
+import java.util.Locale;
+
 /**
  * @author Andrey Pavlenko
  */
 public class MediaStreamInfo {
-	private final int id;
+	private final long id;
 	private final String language;
 	private final String description;
 
-	public MediaStreamInfo(int id, String language, String description) {
+	public MediaStreamInfo(long id, String language, String description) {
 		this.id = id;
-		this.language = language;
+		this.language = (language == null) ? null :
+				language.contains(" + ") ? language : new Locale(language).getDisplayLanguage();
 		this.description = description;
 	}
 
-	public int getId() {
+	public long getId() {
 		return id;
 	}
 
@@ -31,13 +34,13 @@ public class MediaStreamInfo {
 
 	@Override
 	public int hashCode() {
-		return getId();
+		return Long.hashCode(getId());
 	}
 
 	@Override
 	public boolean equals(@Nullable Object obj) {
-		return (obj == this) || ((getClass().isInstance(obj))
-				&& ((MediaStreamInfo) obj).getId() == getId());
+		return (obj == this) ||
+				((getClass().isInstance(obj)) && ((MediaStreamInfo) obj).getId() == getId());
 	}
 
 	@NonNull
@@ -49,7 +52,7 @@ public class MediaStreamInfo {
 		boolean descEmpty = (desc == null) || (desc = desc.trim()).isEmpty();
 
 		if (langEmpty && descEmpty) {
-			return String.valueOf(getId());
+			return "Track " + getId();
 		} else if (langEmpty) {
 			return desc;
 		} else if (descEmpty) {
