@@ -1,5 +1,7 @@
 package me.aap.fermata.media.lib;
 
+import static java.util.Collections.emptyList;
+import static java.util.Collections.emptyMap;
 import static me.aap.utils.async.Completed.completed;
 import static me.aap.utils.async.Completed.completedNull;
 import static me.aap.utils.text.TextUtils.indexOfChar;
@@ -7,6 +9,7 @@ import static me.aap.utils.text.TextUtils.indexOfChar;
 import android.net.Uri;
 
 import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
 
 import java.io.BufferedReader;
 import java.io.IOException;
@@ -15,6 +18,7 @@ import java.io.InputStreamReader;
 import java.io.Reader;
 import java.util.ArrayList;
 import java.util.Collections;
+import java.util.HashMap;
 import java.util.LinkedHashMap;
 import java.util.List;
 import java.util.Map;
@@ -30,6 +34,7 @@ import me.aap.fermata.vfs.m3u.M3uFileSystem;
 import me.aap.utils.app.App;
 import me.aap.utils.async.FutureRef;
 import me.aap.utils.async.FutureSupplier;
+import me.aap.utils.collection.CollectionUtils;
 import me.aap.utils.log.Log;
 import me.aap.utils.text.SharedTextBuilder;
 import me.aap.utils.text.TextUtils;
@@ -50,7 +55,7 @@ public class M3uItem extends BrowsableItemBase {
 			return App.get().execute(M3uItem.this::parse)
 					.ifFail(err -> {
 						Log.e(err, "Failed to load M3U: ", getResource().getName());
-						return new Data(getResource().getName(), "", Collections.emptyList(), null);
+						return new Data(getResource().getName(), "", emptyList(), emptyMap(), null);
 					});
 		}
 	};
@@ -65,6 +70,7 @@ public class M3uItem extends BrowsableItemBase {
 		VirtualFolder dir = m3uFile.getParent().peek();
 		Map<String, M3uGroupItem> groups = new LinkedHashMap<>();
 		List<M3uTrackItem> tracks = new ArrayList<>();
+		Map<String, List<M3uTrackItem>> uriToTrack = new HashMap<>();
 		String idPath = id.substring(getScheme().length());
 		VfsManager vfs = getLib().getVfsManager();
 
@@ -239,6 +245,8 @@ public class M3uItem extends BrowsableItemBase {
 				VirtualResource file = vfs.resolve(l, dir).get(null);
 
 				if (file != null) {
+					M3uTrackItem track;
+
 					if (album == null) album = m3uAlbum;
 					if (artist == null) artist = m3uArtist;
 					if (genre == null) genre = m3uGenre;
@@ -247,9 +255,10 @@ public class M3uItem extends BrowsableItemBase {
 					else if ((logoUrlBase != null) && !logo.contains("://")) logo = logoUrlBase + '/' + logo;
 
 					if (group == null) {
-						tracks.add(createTrack(this, -1, tracks.size(), idPath, file, name, album, artist,
+						track = createTrack(this, -1, tracks.size(), idPath, file, name, album, artist,
 								genre, logo, tvgId, tvgName, duration, type, trackCatchup, trackCatchupDays,
-								trackCatchupSource));
+								trackCatchupSource);
+						tracks.add(track);
 					} else {
 						M3uGroupItem g = groups.get(group);
 
@@ -258,10 +267,19 @@ public class M3uItem extends BrowsableItemBase {
 							groups.put(group, g);
 						}
 
-						g.tracks.add(createTrack(g, g.getGroupId(), g.tracks.size(), idPath, file, name, album,
+						track = createTrack(g, g.getGroupId(), g.tracks.size(), idPath, file, name, album,
 								artist, genre, logo, tvgId, tvgName, duration, type, trackCatchup,
-								trackCatchupDays, trackCatchupSource));
+								trackCatchupDays, trackCatchupSource);
+						g.tracks.add(track);
 					}
+
+					CollectionUtils.compute(uriToTrack, track.getResource().getRid().toString(), (u, t) -> {
+						if (t == null) return Collections.singletonList(track);
+						List<M3uTrackItem> values = new ArrayList<>(t.size() + 1);
+						values.addAll(t);
+						values.add(track);
+						return values;
+					});
 				}
 
 				name = group = album = artist = genre = logo = tvgId = tvgName = null;
@@ -292,7 +310,7 @@ public class M3uItem extends BrowsableItemBase {
 
 		String title = (m3uName == null) ? m3uFile.getName() : m3uName;
 		String subtitle = createSubtitle(ngroups, ntracks);
-		return new Data(title, subtitle, children, cover);
+		return new Data(title, subtitle, children, uriToTrack, cover);
 	}
 
 	@NonNull
@@ -358,7 +376,7 @@ public class M3uItem extends BrowsableItemBase {
 
 	protected M3uGroupItem createGroup(String idPath, String name, int groupId) {
 		SharedTextBuilder tb = SharedTextBuilder.get();
-		tb.append(M3uGroupItem.SCHEME).append(':').append(groupId).append(idPath);
+		tb.append(M3uGroupItem.SCHEME).append(':').append(groupId).append(idPath).append(':').append(name);
 		return new M3uGroupItem(tb.releaseString(), this, name, groupId);
 	}
 
@@ -369,7 +387,7 @@ public class M3uItem extends BrowsableItemBase {
 																		 String catchup, String catchupDays, String catchupSource) {
 		SharedTextBuilder tb = SharedTextBuilder.get();
 		tb.append(M3uTrackItem.SCHEME).append(':').append(groupNumber).append(':')
-				.append(trackNumber).append(idPath);
+				.append(trackNumber).append(idPath).append(':').append(file.getRid());
 		return new M3uTrackItem(tb.releaseString(), parent, trackNumber, file, name, album, artist,
 				genre, logo, tvgId, tvgName, duration, type);
 	}
@@ -410,7 +428,7 @@ public class M3uItem extends BrowsableItemBase {
 
 	public FutureSupplier<? extends M3uTrackItem> getTrack(int id) {
 		return getData().get().map(d -> {
-			for (Item c : d.tracks) {
+			for (Item c : d.items) {
 				if (c instanceof M3uTrackItem) {
 					M3uTrackItem t = (M3uTrackItem) c;
 					if (t.getTrackNumber() == id) return t;
@@ -420,20 +438,46 @@ public class M3uItem extends BrowsableItemBase {
 		});
 	}
 
-	public FutureSupplier<? extends M3uTrackItem> getTrack(int gid, int tid) {
+	public FutureSupplier<? extends M3uTrackItem> getTrack(int gid, int tid, @Nullable String uri) {
+		if (uri != null) {
+			return getData().get().then(d -> {
+				List<M3uTrackItem> tracks = d.uriToTrack.get(uri);
+				if (tracks == null) return getTrack(gid, tid, null).cast();
+				for (M3uTrackItem t : tracks) {
+					if (t.getTrackNumber() == tid) return completed(t);
+				}
+				return completed(tracks.get(0));
+			});
+		}
 		if (gid == -1) return getTrack(tid);
-		return getGroup(gid).map(g -> (g != null) ? g.getTrack(tid) : null);
+		return getGroup(gid, null).map(g -> (g != null) ? g.getTrack(tid) : null);
 	}
 
-	public FutureSupplier<? extends M3uGroupItem> getGroup(int id) {
+	public FutureSupplier<? extends M3uGroupItem> getGroup(int id, @Nullable String name) {
 		return getData().get().map(d -> {
-			for (Item c : d.tracks) {
+			M3uGroupItem group = null;
+
+			for (Item c : d.items) {
 				if (c instanceof M3uGroupItem) {
 					M3uGroupItem g = (M3uGroupItem) c;
-					if (g.getGroupId() == id) return g;
+					if (g.getGroupId() == id) {
+						if ((name == null) || name.equals(g.getName())) return g;
+						group = g;
+						break;
+					}
 				}
 			}
-			return null;
+
+			if (name == null) return null;
+
+			for (Item c : d.items) {
+				if (c instanceof M3uGroupItem) {
+					M3uGroupItem g = (M3uGroupItem) c;
+					if (name.equals(g.getName())) return g;
+				}
+			}
+
+			return group;
 		});
 	}
 
@@ -444,7 +488,7 @@ public class M3uItem extends BrowsableItemBase {
 
 	@Override
 	public FutureSupplier<List<Item>> listChildren() {
-		return getData().get().map(d -> d.tracks);
+		return getData().get().map(d -> d.items);
 	}
 
 	@Override
@@ -465,7 +509,7 @@ public class M3uItem extends BrowsableItemBase {
 			synchronized (lib.cacheLock()) {
 				lib.removeFromCache(this);
 
-				for (Item i : d.tracks) {
+				for (Item i : d.items) {
 					if (i instanceof M3uGroupItem) {
 						for (M3uTrackItem t : ((M3uGroupItem) i).tracks) {
 							lib.removeFromCache(t);
@@ -518,13 +562,15 @@ public class M3uItem extends BrowsableItemBase {
 	protected static final class Data {
 		protected final String name;
 		protected final String subtitle;
-		protected final List<Item> tracks;
+		protected final List<Item> items;
+		protected final Map<String, List<M3uTrackItem>> uriToTrack;
 		protected final String cover;
 
-		Data(String name, String subtitle, List<Item> tracks, String cover) {
+		Data(String name, String subtitle, List<Item> items, Map<String, List<M3uTrackItem>> uriToTrack, String cover) {
 			this.name = name;
 			this.subtitle = subtitle;
-			this.tracks = tracks;
+			this.items = items;
+			this.uriToTrack = uriToTrack;
 			this.cover = cover;
 		}
 	}
