@@ -1,6 +1,10 @@
 package me.aap.fermata.ui.view;
 
 import static java.util.Objects.requireNonNull;
+import static me.aap.fermata.media.pref.MediaPrefs.HW_ACCEL_AUTO;
+import static me.aap.fermata.media.pref.MediaPrefs.HW_ACCEL_DECODING;
+import static me.aap.fermata.media.pref.MediaPrefs.HW_ACCEL_DISABLED;
+import static me.aap.fermata.media.pref.MediaPrefs.HW_ACCEL_FULL;
 import static me.aap.fermata.media.pref.MediaPrefs.SCALE_16_9;
 import static me.aap.fermata.media.pref.MediaPrefs.SCALE_4_3;
 import static me.aap.fermata.media.pref.MediaPrefs.SCALE_BEST;
@@ -16,6 +20,8 @@ import static me.aap.utils.async.Completed.completed;
 import static me.aap.utils.async.Completed.completedVoid;
 
 import android.content.Context;
+
+import androidx.annotation.Nullable;
 
 import java.util.Collections;
 import java.util.List;
@@ -61,6 +67,7 @@ import me.aap.utils.vfs.VirtualResource;
 public class MediaItemMenuHandler implements OverlayMenu.SelectionHandler {
 	private final OverlayMenu menu;
 	private final Item item;
+	@Nullable
 	private final MediaItemView view;
 
 	public MediaItemMenuHandler(OverlayMenu menu, MediaItemView view) {
@@ -87,9 +94,16 @@ public class MediaItemMenuHandler implements OverlayMenu.SelectionHandler {
 		menu.showFuture(this::build);
 	}
 
-	public FutureSupplier<Void> build(OverlayMenu.Builder builder) {
+	@Nullable
+	private MediaEngine getEngine() {
+		return getMainActivity().getMediaServiceBinder().getCurrentEngine();
+	}
+
+	private FutureSupplier<Void> build(OverlayMenu.Builder builder) {
 		FutureSupplier<Void> r;
 		MainActivityDelegate a = getMainActivity();
+
+		if (canDelete()) builder.addItem(R.id.delete, R.drawable.delete, R.string.delete);
 
 		if (item instanceof PlayableItem) {
 			buildPlayableMenu(a, builder, (PlayableItem) item, true);
@@ -103,41 +117,22 @@ public class MediaItemMenuHandler implements OverlayMenu.SelectionHandler {
 		return r.onSuccess(v -> {
 			MediaLibFragment f = a.getActiveMediaLibFragment();
 			if (f != null) f.contributeToContextMenu(builder, this);
-			if (canDelete()) builder.addItem(R.id.delete, R.drawable.delete, R.string.delete);
 			builder.setSelectionHandler(this);
 		});
 	}
 
 	private boolean canDelete() {
+		if (view == null) return false;
 		BrowsableItem p = item.getParent();
 		if ((p == null) || (p.getParent() == null)) return false;
-		return (((item instanceof FileItem) || (item instanceof FolderItem) || (item instanceof M3uItem))
-				&& item.getResource().canDelete());
+		return (
+				((item instanceof FileItem) || (item instanceof FolderItem) || (item instanceof M3uItem)) &&
+						item.getResource().canDelete());
 	}
 
 	protected void buildPlayableMenu(MainActivityDelegate a, OverlayMenu.Builder b, PlayableItem pi,
 																	 boolean initRepeat) {
-		if (pi.isExternal()) return;
-
-		if (pi instanceof StreamItem) {
-			b.addItem(R.id.programme_guide, R.drawable.epg, R.string.programme_guide);
-		}
-
-		if (pi.isFavoriteItem()) {
-			b.addItem(R.id.favorites_remove, R.drawable.favorite_filled, R.string.favorites_remove);
-		} else {
-			b.addItem(R.id.favorites_add, R.drawable.favorite, R.string.favorites_add);
-		}
-
-		if (!(item instanceof StreamItem) && !(item instanceof ArchiveItem)) {
-			if (pi.getPrefs().hasPref(BOOKMARKS)) {
-				b.addItem(R.id.bookmarks, R.drawable.bookmark_filled, R.string.bookmarks)
-						.setFutureSubmenu(this::buildBookmarksMenu);
-			} else {
-				b.addItem(R.id.bookmark_create, R.drawable.bookmark, R.string.create_bookmark)
-						.setSubmenu(this::buildCreateBookmarkMenu);
-			}
-
+		if (!pi.isExternal()) {
 			if (initRepeat) {
 				if (pi.isRepeatItemEnabled()) {
 					b.addItem(R.id.repeat_disable, R.drawable.repeat_filled, R.string.repeat_disable);
@@ -145,30 +140,130 @@ public class MediaItemMenuHandler implements OverlayMenu.SelectionHandler {
 					b.addItem(R.id.repeat_enable, R.drawable.repeat, R.string.repeat);
 				}
 			}
+
+			if (pi instanceof StreamItem) {
+				b.addItem(R.id.programme_guide, R.drawable.epg, R.string.programme_guide);
+			}
+
+			if (pi.isFavoriteItem()) {
+				b.addItem(R.id.favorites_remove, R.drawable.favorite_filled, R.string.favorites_remove);
+			} else {
+				b.addItem(R.id.favorites_add, R.drawable.favorite, R.string.favorites_add);
+			}
+
+			if ((pi.getParent() instanceof Playlist)) {
+				b.addItem(R.id.playlist_remove_item, R.drawable.playlist_remove,
+						R.string.playlist_remove_item);
+			} else {
+				a.addPlaylistMenu(b, completed(Collections.singletonList(pi)));
+			}
+
+			if (!(item instanceof StreamItem) && !(item instanceof ArchiveItem)) {
+				if (pi.getPrefs().hasPref(BOOKMARKS)) {
+					b.addItem(R.id.bookmarks, R.drawable.bookmark_filled, R.string.bookmarks)
+							.setFutureSubmenu(this::buildBookmarksMenu);
+				} else {
+					b.addItem(R.id.bookmark_create, R.drawable.bookmark, R.string.create_bookmark)
+							.setSubmenu(this::buildCreateBookmarkMenu);
+				}
+			}
 		}
 
-		if (pi.isVideo()) {
+		if (pi.isVideo() && addVideoMenu()) {
 			b.addItem(R.id.video, R.drawable.video, R.string.video).setSubmenu(this::buildVideoMenu);
 		}
 
-		if ((pi.getParent() instanceof Playlist)) {
-			b.addItem(R.id.playlist_remove_item, R.drawable.playlist_remove, R.string.playlist_remove_item);
-		} else {
-			a.addPlaylistMenu(b, completed(Collections.singletonList(pi)));
+		if (addSubtitlesMenu()) {
+			b.addItem(R.id.subtitle_prefs, R.drawable.subtitles, R.string.subtitles)
+					.setSubmenu(this::buildSubtitlesMenu);
 		}
 
-		addMediaEngineMenu(a, b);
+		if (addMediaEngMenu()) {
+			b.addItem(R.id.preferred_media_engine, R.drawable.media_engine,
+					R.string.preferred_media_engine).setSubmenu(this::buildVideoEngMenu);
+		}
 	}
 
-	private FutureSupplier<Void> buildBrowsableMenu(MainActivityDelegate a, OverlayMenu.Builder b, BrowsableItem bi) {
-		return bi.getUnsortedChildren().main().then(children -> {
-			boolean hasBookmarks = false;
-			BrowsableItem parent = bi.getParent();
+	protected boolean addVideoMenu() {
+		return true;
+	}
 
-			for (Item c : children) {
-				if ((c instanceof PlayableItem) && ((PlayableItem) c).getPrefs().hasPref(BOOKMARKS)) {
-					hasBookmarks = true;
-					break;
+	protected void buildVideoMenu(OverlayMenu.Builder b) {
+		b.setSelectionHandler(this);
+
+		if (!item.isExternal()) {
+			PlayableItem pi = (PlayableItem) item;
+			if (pi.getPrefs().getWatchedPref()) {
+				b.addItem(R.id.mark_unwatched, R.string.mark_unwatched);
+			} else {
+				b.addItem(R.id.mark_watched, R.string.mark_watched);
+			}
+		}
+
+		if (addAudioMenu()) {
+			b.addItem(R.id.audio_prefs, R.string.audio).setSubmenu(this::buildAudioMenu);
+		}
+
+		b.addItem(R.id.video_scaling, R.string.video_scaling).setSubmenu(this::buildVideoScalingMenu);
+
+		if (getMainActivity().getMediaSessionCallback().getEngineManager().isVlcPlayerSupported()) {
+			b.addItem(R.id.video_scaling, R.string.hw_accel).setSubmenu(this::buildHwAccelMenu);
+		}
+
+		if (!item.isExternal()) {
+			b.addItem(R.id.watched_threshold, R.string.watched_threshold)
+					.setSubmenu(this::buildWatchedThresholdMenu);
+		}
+	}
+
+	protected boolean addAudioMenu() {
+		return item.getLib().getPrefs().getVlcEnabledPref();
+	}
+
+	protected void buildAudioMenu(OverlayMenu.Builder b) {
+		if (!getMainActivity().getMediaSessionCallback().getEngineManager().isVlcPlayerSupported())
+			return;
+		PreferenceSet prefSet = new PreferenceSet();
+		addDelayPrefs(prefSet, item.getPrefs(), MediaLibPrefs.AUDIO_DELAY, R.string.audio_delay, null);
+		prefSet.addToMenu(b, true);
+	}
+
+	protected boolean addSubtitlesMenu() {
+		return (item != null) && item.getPrefs().getSubEnabledPref();
+	}
+
+	protected void buildSubtitlesMenu(OverlayMenu.Builder b) {
+		PreferenceSet prefSet = new PreferenceSet();
+		addDelayPrefs(prefSet, item.getPrefs(), MediaLibPrefs.SUB_DELAY, R.string.subtitle_delay,
+				null);
+		prefSet.addToMenu(b, true);
+	}
+
+	protected boolean addMediaEngMenu() {
+		return (view != null) && getMainActivity().getMediaSessionCallback().getEngineManager()
+				.isAdditionalPlayerSupported();
+	}
+
+	private FutureSupplier<Void> buildBrowsableMenu(MainActivityDelegate a, OverlayMenu.Builder b,
+																									BrowsableItem bi) {
+		return bi.getUnsortedChildren().main().then(children -> {
+			var hasAudio = false;
+			var hasVideo = false;
+			var hasWatched = false;
+			var hasBookmarks = false;
+			var parent = bi.getParent();
+
+			for (var c : children) {
+				if (c instanceof PlayableItem pi) {
+					var prefs = pi.getPrefs();
+					if (pi.isVideo()) {
+						hasVideo = true;
+						if (prefs.getWatchedPref()) hasWatched = true;
+					} else {
+						hasAudio = true;
+					}
+					if (prefs.hasPref(BOOKMARKS)) hasBookmarks = true;
+					if (hasAudio && hasVideo && hasWatched && hasBookmarks) break;
 				}
 			}
 
@@ -176,7 +271,8 @@ public class MediaItemMenuHandler implements OverlayMenu.SelectionHandler {
 				b.addItem(R.id.favorites_add, R.drawable.favorite, R.string.favorites_add);
 
 				if ((bi instanceof Playlist)) {
-					b.addItem(R.id.playlist_remove, R.drawable.playlist_remove, R.string.playlist_remove).setData(bi);
+					b.addItem(R.id.playlist_remove, R.drawable.playlist_remove, R.string.playlist_remove)
+							.setData(bi);
 				}
 			}
 
@@ -185,42 +281,68 @@ public class MediaItemMenuHandler implements OverlayMenu.SelectionHandler {
 			}
 
 			if (hasBookmarks) {
-				b.addItem(R.id.bookmarks, R.drawable.bookmark_filled, R.string.bookmarks).setFutureSubmenu(this::buildBookmarksMenu);
+				b.addItem(R.id.bookmarks, R.drawable.bookmark_filled, R.string.bookmarks)
+						.setFutureSubmenu(this::buildBookmarksMenu);
 			}
 			if (!(bi instanceof Playlist)) {
 				a.addPlaylistMenu(b, () -> bi.getPlayableChildren(true), bi::getName);
 			}
 
+			if (hasVideo) {
+				var addUnwatched = hasWatched;
+				b.addItem(R.id.video, R.drawable.video, R.string.video).setSubmenu(sb -> {
+					sb.setSelectionHandler(this);
+					sb.addItem(R.id.mark_watched, R.string.mark_watched);
+					if (addUnwatched) sb.addItem(R.id.mark_unwatched, R.string.mark_unwatched);
+					sb.addItem(R.id.audio_prefs, R.string.audio).setSubmenu(this::buildAudioPrefsMenu);
+					sb.addItem(R.id.video_scaling, R.string.video_scaling)
+							.setSubmenu(this::buildVideoScalingMenu);
+
+					if (a.getMediaSessionCallback().getEngineManager().isVlcPlayerSupported()) {
+						sb.addItem(R.id.video_scaling, R.string.hw_accel)
+								.setSubmenu(this::buildHwAccelMenu);
+					}
+
+					sb.addItem(R.id.watched_threshold, R.string.watched_threshold)
+							.setSubmenu(this::buildWatchedThresholdMenu);
+				});
+			}
+
+			if (hasVideo || hasAudio) {
+				b.addItem(R.id.subtitle_prefs, R.drawable.subtitles, R.string.subtitles)
+						.setSubmenu(this::buildSubtitlePrefsMenu);
+			}
+
 			b.addItem(R.id.playback_settings, R.drawable.playback_settings, R.string.playback_settings)
 					.setSubmenu(sb -> {
-						sb.withSelectionHandler(this);
+						sb.setSelectionHandler(this);
 						sb.addItem(R.id.watched_threshold, R.string.watched_threshold)
-								.setSubmenu(this::buildWatchedDelayMenu);
+								.setSubmenu(this::buildWatchedThresholdMenu);
 						sb.addItem(R.id.play_next, R.string.play_next_on_completion)
 								.setChecked(bi.getPrefs().getPlayNextPref());
 					});
 
-			addMediaEngineMenu(a, b);
+			if ((hasVideo || hasAudio) && addMediaEngMenu()) {
+				if (hasAudio && hasVideo) {
+					b.addItem(R.id.preferred_media_engine, R.drawable.media_engine,
+							R.string.preferred_media_engine).setSubmenu(sb -> {
+						sb.setSelectionHandler(this);
+						sb.addItem(R.id.preferred_audio_engine, R.string.preferred_audio_engine)
+								.setSubmenu(this::buildAudioEngMenu);
+						sb.addItem(R.id.preferred_video_engine, R.string.preferred_video_engine)
+								.setSubmenu(this::buildVideoEngMenu);
+					});
+				} else if (hasAudio) {
+					b.addItem(R.id.preferred_audio_engine, R.drawable.media_engine,
+							R.string.preferred_audio_engine).setSubmenu(this::buildAudioEngMenu);
+				} else {
+					b.addItem(R.id.preferred_video_engine, R.drawable.media_engine,
+							R.string.preferred_video_engine).setSubmenu(this::buildVideoEngMenu);
+				}
+			}
+
 			return completedVoid();
 		});
-	}
-
-	private void addMediaEngineMenu(MainActivityDelegate a, OverlayMenu.Builder builder) {
-		if (a.getMediaSessionCallback().getEngineManager().isAdditionalPlayerSupported()) {
-			builder.addItem(R.id.preferred_media_engine, R.drawable.media_engine,
-					R.string.preferred_media_engine).setSubmenu(this::buildMediaEngineMenu);
-		}
-	}
-
-	protected void buildMediaEngineMenu(OverlayMenu.Builder b) {
-		b.setSelectionHandler(this);
-
-		if (item instanceof PlayableItem) {
-			buildMediaEngMenu(b, ((PlayableItem) item).isVideo());
-		} else if (item instanceof BrowsableItem) {
-			b.addItem(R.id.preferred_audio_engine, R.string.preferred_audio_engine).setSubmenu(this::buildAudioEngMenu);
-			b.addItem(R.id.preferred_video_engine, R.string.preferred_video_engine).setSubmenu(this::buildVideoEngMenu);
-		}
 	}
 
 	private void buildAudioEngMenu(OverlayMenu.Builder b) {
@@ -234,59 +356,33 @@ public class MediaItemMenuHandler implements OverlayMenu.SelectionHandler {
 	private void buildMediaEngMenu(OverlayMenu.Builder b, boolean video) {
 		MediaPrefs prefs = item.getPrefs();
 		MediaEngineManager mgr = getMainActivity().getMediaSessionCallback().getEngineManager();
-		Pref<IntSupplier> p = video ? MediaPrefs.VIDEO_ENGINE.withInheritance(false)
-				: MediaPrefs.AUDIO_ENGINE.withInheritance(false);
-		int eng = prefs.hasPref(p) ? (video ? prefs.getVideoEnginePref() : prefs.getAudioEnginePref()) : -1;
+		Pref<IntSupplier> p = video ? MediaPrefs.VIDEO_ENGINE.withInheritance(false) :
+				MediaPrefs.AUDIO_ENGINE.withInheritance(false);
+		int eng =
+				prefs.hasPref(p) ? (video ? prefs.getVideoEnginePref() : prefs.getAudioEnginePref()) : -1;
 		b.setSelectionHandler(this);
 
-		OverlayMenuItem i = b.addItem(video ? R.id.preferred_video_engine_default
-				: R.id.preferred_audio_engine_default, null, R.string.by_default);
+		OverlayMenuItem i =
+				b.addItem(video ? R.id.preferred_video_engine_default :
+								R.id.preferred_audio_engine_default,
+						null, R.string.by_default);
 		i.setChecked(eng == -1, true);
-		i = b.addItem(video ? R.id.preferred_video_engine_mp
-				: R.id.preferred_audio_engine_mp, null, R.string.engine_mp_name);
+		i = b.addItem(video ? R.id.preferred_video_engine_mp : R.id.preferred_audio_engine_mp, null,
+				R.string.engine_mp_name);
 		i.setChecked(eng == MediaPrefs.MEDIA_ENG_MP, true);
 
 		if (mgr.isExoPlayerSupported()) {
-			i = b.addItem(video ? R.id.preferred_video_engine_exo
-					: R.id.preferred_audio_engine_exo, null, R.string.engine_exo_name);
+			i = b.addItem(video ? R.id.preferred_video_engine_exo : R.id.preferred_audio_engine_exo,
+					null,
+					R.string.engine_exo_name);
 			i.setChecked(eng == MediaPrefs.MEDIA_ENG_EXO, true);
 		}
 
 		if (mgr.isVlcPlayerSupported()) {
-			i = b.addItem(video ? R.id.preferred_video_engine_vlc
-					: R.id.preferred_audio_engine_vlc, null, R.string.engine_vlc_name);
+			i = b.addItem(video ? R.id.preferred_video_engine_vlc : R.id.preferred_audio_engine_vlc,
+					null,
+					R.string.engine_vlc_name);
 			i.setChecked(eng == MediaPrefs.MEDIA_ENG_VLC, true);
-		}
-	}
-
-	protected void buildVideoMenu(OverlayMenu.Builder b) {
-		b.setSelectionHandler(this);
-
-		if ((item instanceof PlayableItem)) {
-			PlayableItem pi = (PlayableItem) item;
-
-			if (pi.getPrefs().getWatchedPref()) {
-				b.addItem(R.id.mark_unwatched, R.string.mark_unwatched);
-			} else {
-				b.addItem(R.id.mark_watched, R.string.mark_watched);
-			}
-			b.addItem(R.id.watched_threshold, R.string.watched_threshold)
-					.setSubmenu(this::buildWatchedDelayMenu);
-		}
-
-		b.addItem(R.id.video_scaling, R.string.video_scaling).setSubmenu(this::buildVideoScalingMenu);
-
-		if (item.getLib().getPrefs().getVlcEnabledPref()) {
-			if ((item instanceof PlayableItem)) {
-				b.addItem(R.id.audio_delay, R.string.audio_delay).setSubmenu(this::buildAudioDelayMenu);
-
-				if (((PlayableItem) item).getPrefs().getSubEnabledPref()) {
-					b.addItem(R.id.subtitle_delay, R.string.subtitle_delay).setSubmenu(this::buildSubtitleDelayMenu);
-				}
-			} else if ((item instanceof BrowsableItem)) {
-				b.addItem(R.id.audio_prefs, R.string.audio).setSubmenu(this::buildAudioPrefsMenu);
-				b.addItem(R.id.subtitle_prefs, R.string.subtitles).setSubmenu(this::buildSubtitlePrefsMenu);
-			}
 		}
 	}
 
@@ -300,7 +396,8 @@ public class MediaItemMenuHandler implements OverlayMenu.SelectionHandler {
 
 		if (item instanceof PlayableItem) {
 			addBookmarks(b, Collections.singletonList((PlayableItem) item));
-			b.addItem(R.id.bookmark_create, R.string.create_bookmark).setSubmenu(this::buildCreateBookmarkMenu);
+			b.addItem(R.id.bookmark_create, R.string.create_bookmark)
+					.setSubmenu(this::buildCreateBookmarkMenu);
 			return completedVoid();
 		} else if (item instanceof BrowsableItem) {
 			return ((BrowsableItem) item).getPlayableChildren(true).main().then(items -> {
@@ -327,11 +424,10 @@ public class MediaItemMenuHandler implements OverlayMenu.SelectionHandler {
 			Bookmark bm = i.getData();
 			i.getMenu().show(b -> {
 				b.setTitle(i.getTitle());
-				b.addItem(R.id.bookmark_remove, R.string.remove_bookmark)
-						.setHandler(mi -> {
-							((PlayableItem) item).getPrefs().removeBookmark(bm.name, bm.time);
-							return true;
-						});
+				b.addItem(R.id.bookmark_remove, R.string.remove_bookmark).setHandler(mi -> {
+					((PlayableItem) item).getPrefs().removeBookmark(bm.name, bm.time);
+					return true;
+				});
 			});
 		} else {
 			Bookmark bm = i.getData();
@@ -342,26 +438,33 @@ public class MediaItemMenuHandler implements OverlayMenu.SelectionHandler {
 	}
 
 	private void buildVideoScalingMenu(OverlayMenu.Builder b) {
-		int scale = item.getPrefs().hasPref(VIDEO_SCALE, false) ? item.getPrefs().getVideoScalePref() : -1;
+		int scale =
+				item.getPrefs().hasPref(VIDEO_SCALE, false) ? item.getPrefs().getVideoScalePref() : -1;
 		b.addItem(R.id.video_scaling_default, null, R.string.by_default).setChecked(scale == -1, true);
-		b.addItem(R.id.video_scaling_best, null, R.string.video_scaling_best).setChecked(scale == SCALE_BEST, true);
-		b.addItem(R.id.video_scaling_fill, null, R.string.video_scaling_fill).setChecked(scale == SCALE_FILL, true);
-		b.addItem(R.id.video_scaling_orig, null, R.string.video_scaling_orig).setChecked(scale == SCALE_ORIGINAL, true);
-		b.addItem(R.id.video_scaling_4, null, R.string.video_scaling_4).setChecked(scale == SCALE_4_3, true);
-		b.addItem(R.id.video_scaling_16, null, R.string.video_scaling_16).setChecked(scale == SCALE_16_9, true);
+		b.addItem(R.id.video_scaling_best, null, R.string.video_scaling_best)
+				.setChecked(scale == SCALE_BEST, true);
+		b.addItem(R.id.video_scaling_fill, null, R.string.video_scaling_fill)
+				.setChecked(scale == SCALE_FILL, true);
+		b.addItem(R.id.video_scaling_orig, null, R.string.video_scaling_orig)
+				.setChecked(scale == SCALE_ORIGINAL, true);
+		b.addItem(R.id.video_scaling_4, null, R.string.video_scaling_4)
+				.setChecked(scale == SCALE_4_3, true);
+		b.addItem(R.id.video_scaling_16, null, R.string.video_scaling_16)
+				.setChecked(scale == SCALE_16_9, true);
 		b.setSelectionHandler(this);
 	}
 
-	private void buildAudioDelayMenu(OverlayMenu.Builder b) {
-		PreferenceSet prefSet = new PreferenceSet();
-		addDelayPrefs(prefSet, item.getPrefs(), MediaLibPrefs.AUDIO_DELAY, R.string.audio_delay, null);
-		prefSet.addToMenu(b, true);
-	}
-
-	private void buildSubtitleDelayMenu(OverlayMenu.Builder b) {
-		PreferenceSet prefSet = new PreferenceSet();
-		addDelayPrefs(prefSet, item.getPrefs(), MediaLibPrefs.SUB_DELAY, R.string.subtitle_delay, null);
-		prefSet.addToMenu(b, true);
+	private void buildHwAccelMenu(OverlayMenu.Builder b) {
+		int accel = item.getPrefs().getHwAccelPref();
+		b.addItem(R.id.hw_accel_auto, null, R.string.hw_accel_auto)
+				.setChecked(accel == HW_ACCEL_AUTO, true);
+		b.addItem(R.id.hw_accel_full, null, R.string.hw_accel_full)
+				.setChecked(accel == HW_ACCEL_FULL, true);
+		b.addItem(R.id.hw_accel_decoding, null, R.string.hw_accel_decoding)
+				.setChecked(accel == HW_ACCEL_DECODING, true);
+		b.addItem(R.id.hw_accel_disabled, null, R.string.hw_accel_disabled)
+				.setChecked(accel == HW_ACCEL_DISABLED, true);
+		b.setSelectionHandler(this);
 	}
 
 	private void buildAudioPrefsMenu(OverlayMenu.Builder b) {
@@ -378,7 +481,7 @@ public class MediaItemMenuHandler implements OverlayMenu.SelectionHandler {
 		prefSet.addToMenu(b, true);
 	}
 
-	private void buildWatchedDelayMenu(OverlayMenu.Builder b) {
+	private void buildWatchedThresholdMenu(OverlayMenu.Builder b) {
 		PreferenceSet prefSet = new PreferenceSet();
 		prefSet.addIntPref(o -> {
 			o.store = item.getPrefs();
@@ -422,8 +525,7 @@ public class MediaItemMenuHandler implements OverlayMenu.SelectionHandler {
 			f = getMainActivity().getMediaLibFragment(R.id.favorites_fragment);
 			if (f != null) f.reload();
 		} else if (id == R.id.playlist_remove_item) {
-			getMainActivity().removeFromPlaylist(
-					(Playlist) requireNonNull(item.getParent()),
+			getMainActivity().removeFromPlaylist((Playlist) requireNonNull(item.getParent()),
 					Collections.singletonList((PlayableItem) item));
 		} else if (id == R.id.playlist_remove) {
 			Playlist p = i.getData();
@@ -444,7 +546,15 @@ public class MediaItemMenuHandler implements OverlayMenu.SelectionHandler {
 			BrowsableItemPrefs brPrefs = ((BrowsableItem) item).getPrefs();
 			brPrefs.setPlayNextPref(!brPrefs.getPlayNextPref());
 		} else if (id == R.id.mark_watched || id == R.id.mark_unwatched) {
-			((PlayableItem) item).getPrefs().setWatchedPref(id == R.id.mark_watched);
+			boolean watched = id == R.id.mark_watched;
+
+			if (item instanceof BrowsableItem br) {
+				br.getPlayableChildren(false).onSuccess(list -> {
+					for (var pi : list) pi.getPrefs().setWatchedPref(watched);
+				});
+			} else if (item instanceof PlayableItem pi) {
+				pi.getPrefs().setWatchedPref(watched);
+			}
 
 			if (view != null) {
 				view.refresh();
@@ -452,9 +562,11 @@ public class MediaItemMenuHandler implements OverlayMenu.SelectionHandler {
 				ActivityFragment mf = getMainActivity().getActiveFragment();
 				if (mf instanceof MediaLibFragment) ((MediaLibFragment) mf).getAdapter().refresh();
 			}
-		} else if (id == R.id.preferred_audio_engine_default || id == R.id.preferred_video_engine_default) {
-			item.getPrefs().removePref((id == R.id.preferred_audio_engine_default)
-					? MediaPrefs.AUDIO_ENGINE : MediaPrefs.VIDEO_ENGINE);
+		} else if (id == R.id.preferred_audio_engine_default ||
+				id == R.id.preferred_video_engine_default) {
+			item.getPrefs().removePref(
+					(id == R.id.preferred_audio_engine_default) ? MediaPrefs.AUDIO_ENGINE :
+							MediaPrefs.VIDEO_ENGINE);
 		} else if (id == R.id.preferred_audio_engine_mp) {
 			item.getPrefs().setAudioEnginePref(MediaPrefs.MEDIA_ENG_MP);
 		} else if (id == R.id.preferred_video_engine_mp) {
@@ -479,14 +591,22 @@ public class MediaItemMenuHandler implements OverlayMenu.SelectionHandler {
 			item.getPrefs().setVideoScalePref(SCALE_4_3);
 		} else if (id == R.id.video_scaling_16) {
 			item.getPrefs().setVideoScalePref(SCALE_16_9);
+		} else if (id == R.id.hw_accel_auto) {
+			item.getPrefs().setHwAccelPref(HW_ACCEL_AUTO);
+		} else if (id == R.id.hw_accel_full) {
+			item.getPrefs().setHwAccelPref(HW_ACCEL_FULL);
+		} else if (id == R.id.hw_accel_decoding) {
+			item.getPrefs().setHwAccelPref(HW_ACCEL_DECODING);
+		} else if (id == R.id.hw_accel_disabled) {
+			item.getPrefs().setHwAccelPref(HW_ACCEL_DISABLED);
 		} else if (id == R.id.delete) {
-			UiUtils.showQuestion(getContext(), R.string.delete_file_title,
-					R.string.delete_file_question, R.drawable.delete).onSuccess(v -> {
+			UiUtils.showQuestion(getContext(), R.string.delete_file_title, R.string.delete_file_question,
+					R.drawable.delete).onSuccess(v -> {
 				VirtualResource res = item.getResource();
 				res.delete().main().onCompletion((deleted, err) -> {
 					if ((err == null) && deleted) {
 						MainActivityDelegate a = getMainActivity();
-						MediaEngine eng = a.getMediaServiceBinder().getCurrentEngine();
+						MediaEngine eng = getEngine();
 						if ((eng != null) && item.equals(eng.getSource()))
 							a.getMediaSessionCallback().onSkipToNext();
 						ActivityFragment mf = a.getActiveFragment();
@@ -512,7 +632,7 @@ public class MediaItemMenuHandler implements OverlayMenu.SelectionHandler {
 	}
 
 	private void buildCreateBookmarkMenu(OverlayMenu.Builder b) {
-		MediaEngine eng = getMainActivity().getMediaServiceBinder().getCurrentEngine();
+		MediaEngine eng = getEngine();
 
 		if ((eng != null) && item.equals(eng.getSource())) {
 			eng.getPosition().onSuccess(pos -> buildCreateBookmarkMenu(b, (int) (pos / 1000)));
