@@ -5,11 +5,13 @@ import static me.aap.utils.net.http.HttpFileDownloader.AGENT;
 import static me.aap.utils.net.http.HttpFileDownloader.CHARSET;
 import static me.aap.utils.net.http.HttpFileDownloader.ENCODING;
 import static me.aap.utils.net.http.HttpFileDownloader.MAX_AGE;
+import static me.aap.utils.net.http.HttpFileDownloader.RESP_TIMEOUT;
 import static me.aap.utils.text.TextUtils.isInt;
 import static me.aap.utils.text.TextUtils.trim;
 
 import android.content.Context;
 import android.content.SharedPreferences;
+import android.net.Uri;
 
 import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
@@ -103,6 +105,14 @@ public class M3uFile implements VirtualFile {
 		getPrefs().applyStringPref(AGENT, trim(agent));
 	}
 
+	public int getResponseTimeout() {
+		return getPrefs().getIntPref(RESP_TIMEOUT);
+	}
+
+	public void setResponseTimeout(int timeout) {
+		getPrefs().applyIntPref(RESP_TIMEOUT, timeout);
+	}
+
 	@NonNull
 	@Override
 	public Rid getRid() {
@@ -126,6 +136,12 @@ public class M3uFile implements VirtualFile {
 
 	@Override
 	public AsyncInputStream getInputStream() throws IOException {
+		String url = getUrl();
+
+		if (url.startsWith("content://")) {
+			return AsyncInputStream.from(App.get().getContentResolver().openInputStream(Uri.parse(url)));
+		}
+
 		return LocalFileSystem.getInstance().getFile(getLocalFile()).getInputStream();
 	}
 
@@ -139,8 +155,8 @@ public class M3uFile implements VirtualFile {
 
 	@Nullable
 	@Override
-	public RandomAccessChannel getChannel() {
-		return LocalFileSystem.getInstance().getFile(getLocalFile()).getChannel();
+	public RandomAccessChannel getChannel(String mode) {
+		return LocalFileSystem.getInstance().getFile(getLocalFile()).getChannel(mode);
 	}
 
 	@Override
@@ -158,12 +174,8 @@ public class M3uFile implements VirtualFile {
 	@NonNull
 	@Override
 	public String toString() {
-		return getClass().getSimpleName() + " {" +
-				"rid=" + rid +
-				", name='" + getName() + '\'' +
-				", url='" + getUrl() + '\'' +
-				", localFile=" + getLocalFile() +
-				'}';
+		return getClass().getSimpleName() + " {" + "rid=" + rid + ", name='" + getName() + '\'' +
+				", url='" + getUrl() + '\'' + ", localFile=" + getLocalFile() + '}';
 	}
 
 	@NonNull
